Merged PR 6: Spark Notebook Publishing
Added CICD functionality to Spark.  Spark only has an API available instead of real CICD, so it's a small python script to post all of our .R files after they're linted.

Important callout: This is a one-way sync.  Any work done in Spark itself will NOT persist here, and will be overwritten if its in the same files.